# Optimize array_equivalent for NDFrame.equals