@TomAugspurger 2020-07-17
Travis finished. I'm going to selfishly ignore https://github.com/pandas-dev/pandas/pull/35328#discussion_r456589372 so that I can get the RC process rolling, but happy to engage more on followups if I missed something.